requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
click==8.2.1
    # via flask
flask==3.1.1
    # via
    #   -r requirements.in
    #   slackeventsapi
gunicorn==23.0.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==25.0
    # via gunicorn
pyee==11.1.1
    # via slackeventsapi
python-dotenv==1.1.1
    # via -r requirements.in
slack-sdk==3.36.0
    # via -r requirements.in
slackeventsapi==3.0.3
    # via -r requirements.in
typing-extensions==4.14.1
    # via pyee
werkzeug==3.1.3
    # via flask
